+++ pandoc.cabal
@@ -523,8 +523,8 @@ library
                    Text.Pandoc.Readers.Org.Shared,
                    Text.Pandoc.Lua.Filter,
                    Text.Pandoc.Lua.Init,
+                   Text.Pandoc.Lua.Module.Pandoc,
                    Text.Pandoc.Lua.Packages,
-                   Text.Pandoc.Lua.PandocModule,
                    Text.Pandoc.Lua.StackInstances,
                    Text.Pandoc.Lua.Util,
                    Text.Pandoc.CSS,
