+++ src/Text/Pandoc/Lua/Packages.hs
@@ -38,10 +38,11 @@ import Data.IORef (IORef)
 import Foreign.Lua (Lua, NumResults, liftIO)
 import Text.Pandoc.Class (CommonState, readDataFile, runIO, setUserDataDir)
 import Text.Pandoc.MediaBag (MediaBag)
-import Text.Pandoc.Lua.PandocModule (pushPandocModule, pushMediaBagModule)
+import Text.Pandoc.Lua.Module.Pandoc (pushMediaBagModule)
 import Text.Pandoc.Lua.Util (dostring')
 
 import qualified Foreign.Lua as Lua
+import Text.Pandoc.Lua.Module.Pandoc as Pandoc
 
 -- | Parameters used to create lua packages/modules.
 data LuaPackageParams = LuaPackageParams
@@ -72,7 +73,7 @@ pandocPackageSearcher :: LuaPackageParams -> String -> Lua NumResults
 pandocPackageSearcher luaPkgParams pkgName =
   case pkgName of
     "pandoc"          -> let datadir = luaPkgDataDir luaPkgParams
-                         in pushWrappedHsFun (pushPandocModule datadir)
+                         in pushWrappedHsFun (Pandoc.pushModule datadir)
     "pandoc.mediabag" -> let st    = luaPkgCommonState luaPkgParams
                              mbRef = luaPkgMediaBag luaPkgParams
                          in pushWrappedHsFun (pushMediaBagModule st mbRef)
